requirements.txt
coinbase-advanced-py>=1.0.4
python-dotenv>=1.0.0
schedule>=1.2.1
websockets>=12.0
aiohttp>=3.9.3 